requirements.txt
WTForms
Flask
Flask-SQLAlchemy
Flask-Login
flask-Script
Flask-And-Redis
Flask-Mail
Flask-WTF
PyMySQL
